// src/app/directive/widgets-info.enum.ts
export enum WidgetType {
  InertiaCalculator = 'INERTIA_CALCULATOR',
  MotionProfile = 'MOTION_PROFILE',
  TorqueSpeedCurve = 'TORQUE_SPEED_CURVE'
};

export enum WidgetTitle {
  InertiaCalculator = 'Inertia Calculator',
  MotionProfile = 'Motion Profile',
  TorqueSpeedCurve = 'Torque Speed Curve'
};

export enum WidgetIcon {
  InertiaCalculator = 'fa fa-calculator',
  MotionProfile = 'fa fa-chart-area',
  TorqueSpeedCurve = 'fa fa-chart-area'
};